requirements.txt
absl-py==0.13.0
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
certifi==2021.10.8
chardet==4.0.0
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1647850341668/work
decorator==4.4.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.38.1
h5py==3.3.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.6.1
imutils==0.5.4
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
networkx==2.5.1
numpy==1.21.0
onnx==1.7.0
onnx-tf==1.5.0
opencv-python==4.5.2.54
opt-einsum==3.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-msssim==0.2.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.18.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1604304781443/work
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tifffile==2021.7.2
toml==0.10.2
torch==1.5.0
torchsummary==1.5.1
torchtext==0.3.1
torchvision==0.6.0
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
youtube-dl==2021.6.6
zipp==3.5.0
